use f32 for durations and weights

+++ xayn-ai/src/utils.rs
@@ -90,7 +90,7 @@ pub(crate) fn serialize_with_version(data: &impl Serialize, version: u8) -> Resu
 }
 
 /// The number of seconds per day (without leap seconds).
-pub(crate) const SECONDS_PER_DAY: f64 = 86400.;
+pub(crate) const SECONDS_PER_DAY: f32 = 86400.;
 
 /// Gets the current system time depending on the target architecture.
 #[inline]
